Rename pyro.contrib.autoguide.advi to pyro.contrib.autoguide; extend docs
