+++ docs/source/contrib.autoguide.rst
@@ -1,10 +1,52 @@
 Automatic Guide Generation
---------------------------
+==========================
 
-.. automodule:: pyro.contrib.autoguide.advi
+.. automodule:: pyro.contrib.autoguide
+
+AutoGuide
+---------
+.. autoclass:: pyro.contrib.autoguide.AutoGuide
     :members:
     :undoc-members:
     :special-members: __call__
     :show-inheritance:
-    :member-order: bysource
 
+AutoGuideList
+-------------
+.. autoclass:: pyro.contrib.autoguide.AutoGuideList
+    :members:
+    :undoc-members:
+    :special-members: __call__
+    :show-inheritance:
+
+AutoContinuous
+--------------
+.. autoclass:: pyro.contrib.autoguide.AutoContinuous
+    :members:
+    :undoc-members:
+    :special-members: __call__
+    :show-inheritance:
+
+AutoMultivariateNormal
+----------------------
+.. autoclass:: pyro.contrib.autoguide.AutoMultivariateNormal
+    :members:
+    :undoc-members:
+    :special-members: __call__
+    :show-inheritance:
+
+AutoDiagonalNormal
+------------------
+.. autoclass:: pyro.contrib.autoguide.AutoDiagonalNormal
+    :members:
+    :undoc-members:
+    :special-members: __call__
+    :show-inheritance:
+
+AutoDiscreteParallel
+--------------------
+.. autoclass:: pyro.contrib.autoguide.AutoDiscreteParallel
+    :members:
+    :undoc-members:
+    :special-members: __call__
+    :show-inheritance:
