+++ docs/source/contrib.rst
@@ -5,8 +5,8 @@ Contributed Code
     This code makes no guarantee about maintaining backwards compatibility.
 
 .. toctree::
-   :maxdepth: 2
+    :maxdepth: 2
+    :glob:
+    :includehidden:
 
-   contrib.autoguide
-   contrib.gp
-   contrib.named
+    contrib.*
